Move `allocate_shots` function to new module
