+++ src/qibochem/measurement/shot_allocation.py
@@ -2,6 +2,101 @@
 Functions for allocating shots to the Hamiltonian terms
 """
 
+import numpy as np
+
+
+def allocate_shots(grouped_terms, n_shots, method=None, max_shots_per_term=None):
+    """
+    Allocate shots to each group of terms in the Hamiltonian for calculating the expectation value of the Hamiltonian.
+
+    Args:
+        grouped_terms (list): Output of measurement_basis_rotations; list of two-tuples with the first term a
+            :class:`sympy.Expr` and the second the list of corresponding measurement gates (not used here).
+        n_shots (int): Total number of shots to be allocated
+        method (str): How to allocate the shots. The available options are: ``"c"``/``"coefficients"``: ``n_shots`` is
+            distributed based on the relative magnitudes of the term coefficients, ``"u"``/``"uniform"``: ``n_shots``
+            is distributed evenly amongst each term. Default value: ``"c"``.
+        max_shots_per_term (int): Upper limit for the number of shots allocated to an individual group of terms. If not
+            given, will be defined as a fraction (largest coefficient over the sum of all coefficients in the
+            Hamiltonian) of ``n_shots``.
+
+    Returns:
+        list: A list containing the number of shots to be used for each group of Pauli terms respectively.
+    """
+    if method is None:
+        method = "c"
+    if max_shots_per_term is None:
+        # Define based on the fraction of the term group with the largest coefficients w.r.t. sum of all coefficients.
+        term_coefficients = np.array(
+            [
+                sum(abs(coeff) ** (2 / 3) for _t, coeff in expression.as_coefficients_dict().items())
+                for (expression, _) in grouped_terms
+            ],
+            dtype=float,
+        )
+        max_shots_per_term = int(np.ceil(n_shots * (np.max(term_coefficients) / sum(term_coefficients))))
+    max_shots_per_term = min(n_shots, max_shots_per_term)  # Don't let max_shots_per_term > n_shots if manually defined
+
+    n_terms = len(grouped_terms)
+    shot_allocation = np.zeros(n_terms, dtype=int)
+
+    while True:
+        remaining_shots = n_shots - sum(shot_allocation)
+        if not remaining_shots:
+            break
+
+        # In case n_shots is too high s.t. max_shots_per_term is too low
+        # Increase max_shots_per_term, and redo the shot allocation
+        if np.min(shot_allocation) == max_shots_per_term:
+            max_shots_per_term = min(2 * max_shots_per_term, n_shots)
+            shot_allocation = np.zeros(n_terms, dtype=int)
+            continue
+
+        if method in ("c", "coefficients"):
+            # Split shots based on the relative magnitudes of the coefficients of the (group of) Pauli term(s)
+            # and only for terms that haven't reached the upper limit yet
+            term_coefficients = np.array(
+                [
+                    (
+                        sum(abs(coeff) ** (2 / 3) for _t, coeff in expression.as_coefficients_dict().items())
+                        if shots < max_shots_per_term
+                        else 0.0
+                    )
+                    for shots, (expression, _) in zip(shot_allocation, grouped_terms)
+                ],
+                dtype=float,
+            )
+            # Normalise term_coefficients, then get an initial distribution of remaining_shots
+            term_coefficients /= sum(term_coefficients)
+            _shot_allocation = (remaining_shots * term_coefficients).astype(int)
+            # Only keep the terms with >0 shots allocated, renormalise term_coefficients, and distribute again
+            term_coefficients *= _shot_allocation > 0
+            if _shot_allocation.any():
+                term_coefficients /= sum(term_coefficients)
+                _shot_allocation = (remaining_shots * term_coefficients).astype(int)
+            else:
+                # For distributing the remaining few shots, i.e. remaining_shots << n_terms
+                _shot_allocation = np.array(
+                    allocate_shots(grouped_terms, remaining_shots, max_shots_per_term=remaining_shots, method="u")
+                )
+
+        elif method in ("u", "uniform"):
+            # Uniform distribution of shots for every term. Extra shots are randomly distributed
+            _shot_allocation = np.array([remaining_shots // n_terms for _ in range(n_terms)])
+            if not _shot_allocation.any():
+                _shot_allocation = np.zeros(n_terms)
+                _shot_allocation[:remaining_shots] = 1
+                np.random.shuffle(_shot_allocation)
+
+        else:
+            raise NameError("Unknown method!")
+
+        # Add on to the current allocation, and set upper limit to the number of shots for a given term
+        shot_allocation += _shot_allocation.astype(int)
+        shot_allocation = np.clip(shot_allocation, 0, max_shots_per_term)
+
+    return shot_allocation.tolist()
+
 
 def vmsa_shot_allocation(total_shots, n_trial_shots, variance_values):
     """
